requirements.txt
aiohttp==3.7.2
appdirs==1.4.4
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
black==19.3b0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudevents==1.2.0
dapr-dev==0.8.0a0.dev258
dapr-ext-fastapi-dev==0.8.0a0.dev267
dapr-ext-grpc-dev==0.8.0a0.dev258
decorator==4.4.2
deprecation==2.1.0
fastapi==0.61.2
google-api-core==1.23.0
google-auth==1.23.0
googleapis-common-protos==1.52.0
grpcio==1.26.0
h11==0.11.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
multidict==5.0.0
opencensus==0.7.10
opencensus-context==0.1.1
opencensus-ext-grpc==0.7.1
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.2
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
rsa==4.6
six==1.15.0
starlette==0.13.6
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.12.2
wcwidth==0.2.5
yarl==1.6.2
